requirements.txt
Django==4.2.25
channels==4.1.0
channels-redis==4.2.0
celery==5.4.0
django-plotly-dash==2.4.0
plotly==5.24.0
dash==2.17.1
dash-extensions==1.0.14
redis==5.0.4
requests==2.32.3
paho-mqtt==2.1.0
kafka-python==2.0.2
daphne==4.1.2
python-dotenv==1.0.0
